# contracts/openapi/mod-agri-trace.yaml
...
openapi: 3.1.0
info:
  title: SOS Agri Commodity Traceability & Warehouse Receipts API (mod-agri-trace)
  version: 0.1.0
  description: 'National Edition §3. Commodity aggregation, agro-hub warehouse receipts (issued/pledged/redeemed, hash-chained),
    crop traceability provenance chains. Adoptions: Benue (yam), Osun (cocoa), Taraba (tea), Kebbi (rice), Kano (grains).'
paths:
  /agri/v1/farmers:
    post:
      tags:
      - registry
      summary: Register Farmer
      operationId: register_farmer_agri_v1_farmers_post
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmerRegistration'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farmer'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    get:
      tags:
      - registry
      summary: List Farmers
      operationId: list_farmers_agri_v1_farmers_get
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Farmer'
                title: Response List Farmers Agri V1 Farmers Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /agri/v1/farmers/{farmer_id}:
    get:
      tags:
      - registry
      summary: Get Farmer
      operationId: get_farmer_agri_v1_farmers__farmer_id__get
      parameters:
      - name: farmer_id
        in: path
        required: true
        schema:
          type: string
          title: Farmer Id
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farmer'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /agri/v1/commodities:
    get:
      tags:
      - catalog
      summary: Commodities
      description: 'Per-state commodity catalog (Benue yam, Osun cocoa, Taraba tea,

        Kebbi rice, Kano sorghum/maize; generic fallback otherwise).'
      operationId: commodities_agri_v1_commodities_get
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /agri/v1/lots:
    post:
      tags:
      - aggregation
      summary: Intake Lot
      operationId: intake_lot_agri_v1_lots_post
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LotIntake'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lot'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    get:
      tags:
      - aggregation
      summary: List Lots
      operationId: list_lots_agri_v1_lots_get
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lot'
                title: Response List Lots Agri V1 Lots Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /agri/v1/warehouses:
    post:
      tags:
      - warehouse
      summary: Register Warehouse
      operationId: register_warehouse_agri_v1_warehouses_post
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseRegistration'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    get:
      tags:
      - warehouse
      summary: List Warehouses
      operationId: list_warehouses_agri_v1_warehouses_get
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'
                title: Response List Warehouses Agri V1 Warehouses Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /agri/v1/receipts:
    post:
      tags:
      - warehouse-receipts
      summary: Issue Receipt
      operationId: issue_receipt_agri_v1_receipts_post
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptIssueRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseReceipt'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    get:
      tags:
      - warehouse-receipts
      summary: List Receipts
      operationId: list_receipts_agri_v1_receipts_get
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WarehouseReceipt'
                title: Response List Receipts Agri V1 Receipts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /agri/v1/receipts/{receipt_id}:
    get:
      tags:
      - warehouse-receipts
      summary: Get Receipt
      operationId: get_receipt_agri_v1_receipts__receipt_id__get
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          title: Receipt Id
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseReceipt'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /agri/v1/receipts/{receipt_id}/pledge:
    post:
      tags:
      - warehouse-receipts
      summary: Pledge Receipt
      description: Pledge a WR as loan collateral; emits a settlement intent event.
      operationId: pledge_receipt_agri_v1_receipts__receipt_id__pledge_post
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          title: Receipt Id
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PledgeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseReceipt'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
  /agri/v1/receipt-chain/verify:
    get:
      tags:
      - warehouse-receipts
      summary: Verify Receipt Chain
      operationId: verify_receipt_chain_agri_v1_receipt_chain_verify_get
      parameters:
      - name: x-state-tenant
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-State-Tenant
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    FarmerRegistration:
      properties:
        name:
          type: string
          title: Name
        kyc_ref:
          type: string
          title: Kyc Ref
          description: KYC reference from mod-kyc-kyb
        lga:
          type: string
          title: Lga
          default: ''
        phone:
          type: string
          title: Phone
          default: ''
      type: object
      required:
      - name
      - kyc_ref
      title: FarmerRegistration
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReceiptIssueRequest:
      properties:
        warehouse_id:
          type: string
          title: Warehouse Id
        lot_id:
          type: string
          title: Lot Id
        storage_fees_kobo:
          type: integer
          minimum: 0.0
          title: Storage Fees Kobo
      type: object
      required:
      - warehouse_id
      - lot_id
      - storage_fees_kobo
      title: ReceiptIssueRequest
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: https://api.{state}.gov.ng/sos
  description: Per-state sovereign data plane (state = lagos|ogun|osun|benue|nasarawa|taraba)
...